moved notes directory into scripts and updated imports to reflect
