+++ api/notes.json
@@ -13,6 +13,13 @@
       "criminalId": 27,
       "timestamp": 1609950991527,
       "id": 3
+    },
+    {
+      "author": "Jack Thornton",
+      "text": "He was going at least 20 over the speed limit, when he hit me.",
+      "criminalId": 94,
+      "timestamp": 1609953001113,
+      "id": 4
     }
   ]
 }
