+++ scripts/main.js
@@ -1,9 +1,9 @@
 import { CriminalList } from './criminals/criminalList.js';
 import { ConvictionSelect } from './convictions/convictionSelect.js';
 import { OfficerSelect } from './officers/OfficerSelect.js';
-import { NoteForm } from '../notes/NoteForm.js';
-import { ShowNoteButton } from '../notes/ShowNotesButton.js';
-import '../notes/NoteList.js';
+import { NoteForm } from './notes/NoteForm.js';
+import { ShowNoteButton } from './notes/ShowNotesButton.js';
+import './notes/NoteList.js';
 import './alibis/associateAlibiDisplay.js';
 // import { witnessButton } from './witnesses/witnesses.js';
 /*
