+++ scripts/notes/Note.js
