+++ scripts/notes/NoteProvider.js
