+++ scripts/notes/ShowNotesButton.js
